+++ examples/log_async.rs
@@ -30,12 +30,12 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Create a new context with properties.
     let log_context = LogContext::new()
-        .inherited_record("thread_name", "first_future")
-        .local_record("user_id", "12345");
+        .with_inherited_record("thread_name", "first_future")
+        .with_local_record("user_id", "12345");
     let first_future = async move {
         log::info!("Logging in");
         // Create a nested context with additional properties
-        let log_context = LogContext::new().local_record(
+        let log_context = LogContext::new().with_local_record(
             "action",
             LogValue::serde(Operation {
                 action: "login".to_string(),
@@ -55,11 +55,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     .in_log_context(log_context);
 
     let log_context = LogContext::new()
-        .inherited_record("thread_name", "second_future")
-        .local_record("name", "Alice")
-        .local_record("age", 25)
-        .local_record("married", true)
-        .local_record("email", "alice@example.com");
+        .with_inherited_record("thread_name", "second_future")
+        .with_local_record("name", "Alice")
+        .with_local_record("age", 25)
+        .with_local_record("married", true)
+        .with_local_record("email", "alice@example.com");
     let second_future = async move {
         tokio::task::yield_now().await;
 
@@ -70,10 +70,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     .in_log_context(log_context);
 
     let log_context = LogContext::new()
-        .inherited_record("thread_name", "third_future")
-        .local_record("name", "Bob")
-        .local_record("age", 30)
-        .local_record("email", "bob@example.com");
+        .with_inherited_record("thread_name", "third_future")
+        .with_local_record("name", "Bob")
+        .with_local_record("age", 30)
+        .with_local_record("email", "bob@example.com");
     let third_future = tokio::spawn(
         async move {
             tokio::task::yield_now().await;
@@ -97,9 +97,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     res?;
 
     let context = LogContext::new()
-        .local_record("name", "Charlie")
-        .local_record("age", 35)
-        .local_record("email", "charlie@example.com");
+        .with_local_record("name", "Charlie")
+        .with_local_record("age", 35)
+        .with_local_record("email", "charlie@example.com");
 
     let _guard = LogScope::enter(context);